easy_scraper/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
idna==2.10
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
